% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_vis.R
\name{gg_vis_point}
\alias{gg_vis_point}
\title{Creating a ggplot based on a dataset}
\usage{
gg_vis_point(data)
}
\arguments{
\item{data}{a dataset}
}
\value{
A ggplot based on the selected Variables. Either a Scatter- or Lineplot
}
\description{
Creating a ggplot based on a dataset
}
\examples{
gg_vis_point(your_data)
gg_vis_line(mtcars)

}
