% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotvis-package.R
\docType{package}
\name{ggplotvis-package}
\alias{ggplotvis}
\alias{ggplotvis-package}
\title{ggplotvis: Creating a ggplot based on a dataset}
\description{
A ggplot based on the selected Variables, based on a dataset that can be freely choosen
}
\author{
\strong{Maintainer}: Tom Jodl \email{tom.jodl@outlook.com}

}
\keyword{internal}
